requirements.txt
# =========================================================
# Adaptive Stealth Network — Python Dependencies
# =========================================================

# Backend (FastAPI)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0

# Scheduler
apscheduler>=3.10.0

# Controller
httpx>=0.27.0
pyyaml>=6.0

# Logging
rich>=13.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx  # needed for TestClient
